requirements.txt
pyyaml>=6.0
requests>=2.28.0
yt-dlp>=2024.0.0
